[refactor] 프로젝트 구조 변경
- src 아래에 있는 컴포넌트들을 src/components 로 이동
